Add debug features
These features allow us to debug the interpreter, either by dumping the
bytecode or tracing each instruction as it executes.

debug_dump_disassembly is checked using the `cfg!` macro rather than the
attribute so that we don't need to add cfg attributes all over the place
to get the warnings to go away (for example, on the definition of the
VM). The optimizer should eliminate any code from the if statement.