Refactor lexer to use token spans
Previously, the lexer would produce runtime values for tokens such as
strings and numbers. These runtime values would be stored as data on the
TokenType enum. This caused several problems.

First, this obscured the actual source text associated with the token,
making error messages less descriptive. This means that error messages
would for example always display "1.0" for a numeric value of 1, even if
the literal in the source code is "00001.000000".

Second, without actual context for each token in the source code string,
we wouldn't be able to do nice error message stuff like displaying the
line with the problematic span underlined.

Third, since the runtime values are already produced leading into the
compiler, some extra work needs to be done to make sure the values
aren't copied unnecessarily. While this is possible, the code is a bit
ugly.

Now each token contains a range representing its span in the source
string. This solves the above three problems.

I also started running clippy on test code, and modified other files
accordingly.